+++ dessia_common/workflow/core.py
@@ -19,15 +19,14 @@ from dessia_common.core import DessiaObject
 
 from dessia_common.schemas.core import (get_schema, FAILED_ATTRIBUTE_PARSING, EMPTY_PARSED_ATTRIBUTE,
                                         serialize_annotation, pretty_annotation, UNDEFINED, Schema, SchemaAttribute)
-from dessia_common.utils.types import recursive_type, typematch, is_sequence, is_file_or_file_sequence
+from dessia_common.utils.types import recursive_type, typematch, is_file_or_file_sequence
 from dessia_common.utils.copy import deepcopy_value
 from dessia_common.utils.diff import choose_hash
-from dessia_common.utils.helpers import prettyname
+from dessia_common.utils.helpers import prettyname, is_sequence
 from dessia_common.typings import JsonSerializable, ViewType
 from dessia_common.files import StringFile, BinaryFile
 from dessia_common.displays import DisplaySetting
-from dessia_common.breakdown import ExtractionError
-from dessia_common.errors import SerializationError
+from dessia_common.errors import SerializationError, ExtractionError
 from dessia_common.warnings import SerializationWarning
 from dessia_common.exports import ExportFormat, MarkdownWriter
 import dessia_common.templates
@@ -1899,26 +1898,13 @@ class WorkflowRun(WorkflowState):
 
         Concatenate WorkflowState display_settings and inserting Workflow ones.
         """
-        workflow_settings = [display_setting for display_setting in self.workflow.display_settings()
-                             if display_setting.selector != "Documentation"]
         block_settings = self.workflow.blocks_display_settings
 
         displays_by_default = [s.load_by_default for s in block_settings]
         documentation = DisplaySetting(selector="Documentation", type_="markdown", method="to_markdown",
                                        load_by_default=True)
         documentation.load_by_default = not any(displays_by_default)
-
-        workflow_settings_to_keep = [documentation]
-        for settings in workflow_settings:
-            # Update workflow settings
-            settings.compose("workflow")
-
-            if settings.selector == "Workflow":
-                settings.load_by_default = False
-
-            if settings.selector != "Tasks":
-                workflow_settings_to_keep.append(settings)
-        return workflow_settings_to_keep + block_settings
+        return [documentation] + block_settings
 
     def method_dict(self, method_name: str = None, method_jsonschema=None):
         """ Get run again default dict. """